% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS.R
\name{ols}
\alias{ols}
\title{Ordinary least squares (and heteroskedastic‐robust SEs)}
\usage{
ols(Y, X, se = TRUE)
}
\arguments{
\item{Y}{numeric response}

\item{X}{numeric design matrix}

\item{se}{logical; return SEs?}
}
\value{
list(coef, vcov, sXX) or list(coef, sXX)
}
\description{
Ordinary least squares (and heteroskedastic‐robust SEs)
}
